chore: prop to local var
