+++ packages/playwright/src/common/testType.ts
@@ -21,6 +21,7 @@ import { currentTestInfo, currentlyLoadingFileSuite, setCurrentlyLoadingFileSuit
 import { Suite, TestCase } from './test';
 import { expect } from '../matchers/expect';
 import { wrapFunctionWithLocation } from '../transform/transform';
+import { validateAnnotation } from '../util';
 
 import type { FixturesWithLocation } from './config';
 import type { Fixtures, TestAnnotation, TestDetails, TestStepInfo, TestType } from '../../types/test';
@@ -307,40 +308,6 @@ function throwIfRunningInsideJest() {
         `See https://playwright.dev/docs/intro for more information about Playwright Test.`,
     );
   }
-}
-
-export function validateAnnotation(annotation: any): TestAnnotation {
-  if (typeof annotation !== 'object' || annotation === null || Array.isArray(annotation))
-    throw new Error(`Annotation must be an object, received: ${typeof annotation}`);
-
-  if (!('type' in annotation) || annotation.type === null || annotation.type === undefined)
-    throw new Error('Annotation must have a "type" property');
-
-  if (typeof annotation.type !== 'string')
-    throw new Error(`Annotation type must be a string, received: ${typeof annotation.type}`);
-
-  let description = annotation.description;
-  if (description !== undefined && description !== null && typeof description !== 'string')
-    description = JSON.stringify(description);
-
-  if (annotation.location !== undefined) {
-    if (typeof annotation.location !== 'object' || annotation.location === null || Array.isArray(annotation.location))
-      throw new Error(`Annotation location must be an object, received: ${typeof annotation.location}`);
-    if (typeof annotation.location.file !== 'string')
-      throw new Error(`Annotation location.file must be a string, received: ${typeof annotation.location.file}`);
-    if (typeof annotation.location.line !== 'number')
-      throw new Error(`Annotation location.line must be a number, received: ${typeof annotation.location.line}`);
-    if (typeof annotation.location.column !== 'number')
-      throw new Error(`Annotation location.column must be a number, received: ${typeof annotation.location.column}`);
-  }
-
-  const result: TestAnnotation = { type: annotation.type };
-  if (description !== undefined)
-    result.description = description;
-  if (annotation.location)
-    result.location = annotation.location;
-
-  return result;
 }
 
 function validateTestDetails(details: TestDetails, location: Location) {
